opsplisen externalapiservice
